更新了SpiDrv驱动代码

+++ drivers/Proximeter/PMW3901/PMW3901.cpp
@@ -45,18 +45,18 @@ using namespace ymd::drivers;
 using Error = PMW3901::Error;
 
 Result<void, Error> PMW3901::write_reg(const uint8_t command, const uint8_t data){
-    return Result<void, Error>(spi_drv_.write_single<uint8_t>(command | 0x80, CONT)
+    return Result<void, Error>(spi_drv_.write_single<uint8_t>(uint8_t(command | 0x80), CONT)
     | spi_drv_.write_single<uint8_t>(data));
 }
 
 
 Result<void, Error> PMW3901::read_reg(const uint8_t command, uint8_t & data){
-    return Result<void, Error>(spi_drv_.write_single<uint8_t>(command & 0x7f, CONT)
+    return Result<void, Error>(spi_drv_.write_single<uint8_t>(uint8_t(command & 0x7f), CONT)
     | spi_drv_.read_single<uint8_t>(data));
 }
 
 Result<void, Error> PMW3901::read_burst(const uint8_t command, uint8_t * data, const size_t len){
-    return Result<void, Error>(spi_drv_.write_single<uint8_t>(command & 0x7f, CONT)
+    return Result<void, Error>(spi_drv_.write_single<uint8_t>(uint8_t(command & 0x7f), CONT)
     | spi_drv_.read_burst<uint8_t>(data, len));
 }
